requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ca-certificates=2019.10.16=0
certifi=2019.9.11=py37_0
cycler=0.10.0=pypi_0
decorator=4.4.1=pypi_0
echonet=1.0=dev_0
imageio=2.6.1=pypi_0
joblib=0.14.0=pypi_0
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
matplotlib=3.1.1=pypi_0
ncurses=6.1=he6710b0_1
networkx=2.4=pypi_0
numpy=1.17.4=pypi_0
opencv-python=4.1.1.26=pypi_0
openssl=1.0.2t=h7b6447c_1
pillow=6.2.1=pypi_0
pip=19.3.1=py37_0
pyparsing=2.4.5=pypi_0
python=3.7.0=h6e4f718_3
python-dateutil=2.8.1=pypi_0
pywavelets=1.1.1=pypi_0
readline=7.0=h7b6447c_5
scikit-image=0.16.2=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.3.2=pypi_0
setuptools=41.6.0=py37_0
six=1.13.0=pypi_0
sklearn=0.0=pypi_0
sqlite=3.30.1=h7b6447c_0
tk=8.6.8=hbc83047_0
torch=1.3.1=pypi_0
torchvision=0.4.2=pypi_0
tqdm=4.38.0=pypi_0
wheel=0.33.6=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
